.ipynb_checkpoints/requirements-checkpoint.txt
# Core dependencies
requests
pandas
numpy
python-dotenv

# Web scraping
beautifulsoup4
lxml

# AI and NLP
google-generativeai

# Text-to-Speech and Translation
gtts
googletrans

# Web Framework and API
fastapi
uvicorn
streamlit
pydantic

# Visualization
plotly

# Utils
python-multipart